PC Interaction works! (Read Comment for instructions)
Instructions: Go to my nuclear testing scene (that's just what I call my test scenes in every project), and there you can test out the PC controller.
Interact is mapped to left click. The actual character controller prefab is one I found in the samples of the URP package. It is really basic, but works for now. It doesn't even get affected by gravity.
One last thing: playing will cause a lot of compiler errors to get thrown. That's because of the game dialog script attached to the player. That's needed for the audio to properly play and get tracked, but it also handles a lot of the other systems, which I don't have in that scene. So, it throws errors bc it can't find a lot of the core game systems, but that's fine. My scene is just to test the interaction.